refactor(chain-worker): share L1 block refs MMR prefill
