+++ bin/strata/src/genesis.rs
@@ -3,10 +3,10 @@
 use std::collections::BTreeMap;
 
 use anyhow::Result;
-use strata_db_types::{MmrId, ol_block::BlockStatus, ol_state_index::IndexingWrites};
+use strata_chain_worker::prefill_l1_block_refs_mmr_blocking;
+use strata_db_types::{ol_block::BlockStatus, ol_state_index::IndexingWrites};
 use strata_ol_genesis::{GenesisArtifacts, build_genesis_artifacts};
 use strata_ol_params::OLParams;
-use strata_ol_state_types::MMR_SENTINEL_DUMMY_LEAF_HASH;
 use strata_primitives::OLBlockCommitment;
 use strata_storage::NodeStorage;
 use tracing::{info, instrument};
@@ -27,7 +27,10 @@ pub(crate) fn init_ol_genesis(
     } = build_genesis_artifacts(ol_params)?;
     let genesis_blkid = *commitment.blkid();
 
-    prefill_l1_block_refs_mmr(ol_params, storage)?;
+    prefill_l1_block_refs_mmr_blocking(
+        storage.mmr_index(),
+        ol_params.last_l1_block.height() as u64,
+    )?;
 
     // Seed epoch-0 indexing with all genesis accounts as created accounts.
     // Genesis epoch is finalized at boot, so its commitment is known here;
@@ -67,16 +70,4 @@ pub(crate) fn init_ol_genesis(
 
     info!(%genesis_blkid, slot = 0, "OL genesis initialization complete");
     Ok(commitment)
-}
-
-fn prefill_l1_block_refs_mmr(ol_params: &OLParams, storage: &NodeStorage) -> Result<()> {
-    let target_count = ol_params.last_l1_block.height() as u64 + 1;
-    let handle = storage.mmr_index().get_handle(MmrId::L1BlockRefs);
-    let current_count = handle.get_leaf_count_blocking()?;
-
-    for _ in current_count..target_count {
-        handle.append_leaf_blocking(MMR_SENTINEL_DUMMY_LEAF_HASH)?;
-    }
-
-    Ok(())
 }
