+++ crates/chain-worker/src/lib.rs
@@ -18,6 +18,7 @@ mod context;
 mod errors;
 mod handle;
 mod message;
+mod mmr_prefill;
 mod output;
 mod service;
 mod state;
@@ -30,6 +31,7 @@ pub use context::ChainWorkerContextImpl;
 pub use errors::{WorkerError, WorkerResult};
 pub use handle::ChainWorkerHandle;
 pub use message::ChainWorkerMessage;
+pub use mmr_prefill::{prefill_l1_block_refs_mmr, prefill_l1_block_refs_mmr_blocking};
 pub use output::OLBlockExecutionOutput;
 pub use service::{ChainWorkerService, ChainWorkerStatus, start_chain_worker_service_from_ctx};
 pub use state::ChainWorkerServiceState;
